// src/app/pages/cortes/components/cortes-table/cortes.interface.ts
export interface CortesInterface {
   idcorte?: number;
   montoInicial?: number;
   montoFinal?: number;
   ganancia?: number;
   inicia_idempleado?: number;
   finaliza_idempleado?: number;
   fechaInicia?: string;
   horaInicia?: string;
   fechaFInaliza?: string;
   horaFinaliza?: string;
   baja?: boolean;
   modified_at?: string;
   created_by?: number;
   created_at?: string;
}